refact : UserTest 클래스로 리팩토링
